--- easier/examples/.gitignore
@@ -1,5 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-rectangular_triangle_*.npz
-rectangular_triangle_*.hdf5
