--- easier/examples/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-from .models import Poisson
