+++ tests/core/test_jit.py
@@ -13,13 +13,15 @@ import h5py
 import easier as esr
 from easier.core.jit import EasierTracer
 from easier.core.passes.utils import OrderedSet
-from easier.examples import Poisson
 from easier.core.runtime.dist_env import DummyDistEnv
 
 from tests.utils import \
     torchrun_singlenode, get_random_str, \
-    mpi_e2e, mpirun_singlenode, when_ngpus_ge_2
+    mpi_e2e, mpirun_singlenode, when_ngpus_ge_2, \
+    import_poisson, MESH, POISSON
 from tests.core.utils import multi_stage_zero_length_partition
+
+Poisson = import_poisson()
 
 
 class Model(esr.Module):
